+++ test/Microsoft.Azure.Devices.Edge.Test/Device.cs
@@ -1,53 +1,54 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
-{
-    using System;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Devices.Edge.Test.Common;
-    using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
-    using Microsoft.Azure.Devices.Edge.Test.Common.Config;
-    using Microsoft.Azure.Devices.Edge.Test.Helpers;
-    using Microsoft.Azure.Devices.Edge.Util;
-    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
-    using NUnit.Framework;
 
-    [EndToEnd]
-    class Device : SasManualProvisioningFixture
-    {
-        [Test]
-        [Category("CentOsSafe")]
-        public async Task QuickstartCerts()
-        {
-            CancellationToken token = this.TestToken;
+// namespace Microsoft.Azure.Devices.Edge.Test
+// {
+//    using System;
+//    using System.Threading;
+//    using System.Threading.Tasks;
+//    using Microsoft.Azure.Devices.Edge.Test.Common;
+//    using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
+//    using Microsoft.Azure.Devices.Edge.Test.Common.Config;
+//    using Microsoft.Azure.Devices.Edge.Test.Helpers;
+//    using Microsoft.Azure.Devices.Edge.Util;
+//    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
+//    using NUnit.Framework;
 
-            await this.runtime.DeployConfigurationAsync(token);
+// [EndToEnd]
+//    class Device : SasManualProvisioningFixture
+//    {
+//        [Test]
+//        [Category("CentOsSafe")]
+//        public async Task QuickstartCerts()
+//        {
+//            CancellationToken token = this.TestToken;
 
-            string leafDeviceId = DeviceId.Current.Generate();
+// await this.runtime.DeployConfigurationAsync(token);
 
-            var leaf = await LeafDevice.CreateAsync(
-                leafDeviceId,
-                Protocol.Amqp,
-                AuthenticationType.Sas,
-                Option.None<string>(),
-                false,
-                CertificateAuthority.GetQuickstart(),
-                this.iotHub,
-                token,
-                Option.None<string>());
+// string leafDeviceId = DeviceId.Current.Generate();
 
-            await TryFinally.DoAsync(
-                async () =>
-                {
-                    DateTime seekTime = DateTime.Now;
-                    await leaf.SendEventAsync(token);
-                    await leaf.WaitForEventsReceivedAsync(seekTime, token);
-                    await leaf.InvokeDirectMethodAsync(token);
-                },
-                async () =>
-                {
-                    await leaf.DeleteIdentityAsync(token);
-                });
-        }
-    }
-}
+// var leaf = await LeafDevice.CreateAsync(
+//                leafDeviceId,
+//                Protocol.Amqp,
+//                AuthenticationType.Sas,
+//                Option.None<string>(),
+//                false,
+//                CertificateAuthority.GetQuickstart(),
+//                this.iotHub,
+//                token,
+//                Option.None<string>());
+
+// await TryFinally.DoAsync(
+//                async () =>
+//                {
+//                    DateTime seekTime = DateTime.Now;
+//                    await leaf.SendEventAsync(token);
+//                    await leaf.WaitForEventsReceivedAsync(seekTime, token);
+//                    await leaf.InvokeDirectMethodAsync(token);
+//                },
+//                async () =>
+//                {
+//                    await leaf.DeleteIdentityAsync(token);
+//                });
+//        }
+//    }
+// }
