+++ test/Microsoft.Azure.Devices.Edge.Test/Module.cs
@@ -1,182 +1,183 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
-{
-    using System;
-    using System.Collections.Generic;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Devices.Edge.Test.Common;
-    using Microsoft.Azure.Devices.Edge.Test.Common.Config;
-    using Microsoft.Azure.Devices.Edge.Test.Helpers;
-    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
-    using NUnit.Framework;
 
-    [EndToEnd]
-    public class Module : SasManualProvisioningFixture
-    {
-        const string SensorName = "tempSensor";
-        const string DefaultSensorImage = "mcr.microsoft.com/azureiotedge-simulated-temperature-sensor:1.0";
+// namespace Microsoft.Azure.Devices.Edge.Test
+// {
+//    using System;
+//    using System.Collections.Generic;
+//    using System.Threading;
+//    using System.Threading.Tasks;
+//    using Microsoft.Azure.Devices.Edge.Test.Common;
+//    using Microsoft.Azure.Devices.Edge.Test.Common.Config;
+//    using Microsoft.Azure.Devices.Edge.Test.Helpers;
+//    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
+//    using NUnit.Framework;
 
-        [Test]
-        [Category("CentOsSafe")]
-        public async Task TempSensor()
-        {
-            string sensorImage = Context.Current.TempSensorImage.GetOrElse(DefaultSensorImage);
-            CancellationToken token = this.TestToken;
+// [EndToEnd]
+//    public class Module : SasManualProvisioningFixture
+//    {
+//        const string SensorName = "tempSensor";
+//        const string DefaultSensorImage = "mcr.microsoft.com/azureiotedge-simulated-temperature-sensor:1.0";
 
-            EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
-                builder =>
-                {
-                    builder.AddModule(SensorName, sensorImage)
-                        .WithEnvironment(new[] { ("MessageCount", "1") });
-                },
-                token);
+// [Test]
+//        [Category("CentOsSafe")]
+//        public async Task TempSensor()
+//        {
+//            string sensorImage = Context.Current.TempSensorImage.GetOrElse(DefaultSensorImage);
+//            CancellationToken token = this.TestToken;
 
-            EdgeModule sensor = deployment.Modules[SensorName];
-            await sensor.WaitForEventsReceivedAsync(deployment.StartTime, token);
+// EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
+//                builder =>
+//                {
+//                    builder.AddModule(SensorName, sensorImage)
+//                        .WithEnvironment(new[] { ("MessageCount", "1") });
+//                },
+//                token);
 
-            await sensor.UpdateDesiredPropertiesAsync(
-                new
-                {
-                    properties = new
-                    {
-                        desired = new
-                        {
-                            SendData = true,
-                            SendInterval = 10
-                        }
-                    }
-                },
-                token);
-            await sensor.WaitForReportedPropertyUpdatesAsync(
-                new
-                {
-                    properties = new
-                    {
-                        reported = new
-                        {
-                            SendData = true,
-                            SendInterval = 10
-                        }
-                    }
-                },
-                token);
-        }
+// EdgeModule sensor = deployment.Modules[SensorName];
+//            await sensor.WaitForEventsReceivedAsync(deployment.StartTime, token);
 
-        [Test]
-        [Category("CentOsSafe")]
-        public async Task TempFilter()
-        {
-            const string filterName = "tempFilter";
+// await sensor.UpdateDesiredPropertiesAsync(
+//                new
+//                {
+//                    properties = new
+//                    {
+//                        desired = new
+//                        {
+//                            SendData = true,
+//                            SendInterval = 10
+//                        }
+//                    }
+//                },
+//                token);
+//            await sensor.WaitForReportedPropertyUpdatesAsync(
+//                new
+//                {
+//                    properties = new
+//                    {
+//                        reported = new
+//                        {
+//                            SendData = true,
+//                            SendInterval = 10
+//                        }
+//                    }
+//                },
+//                token);
+//        }
 
-            string filterImage = Context.Current.TempFilterImage.Expect(
-                () => new ArgumentException("tempFilterImage parameter is required for TempFilter test"));
-            string sensorImage = Context.Current.TempSensorImage.Expect(
-                () => new ArgumentException("tempSensorImage parameter is required for TempFilter test"));
+// [Test]
+//        [Category("CentOsSafe")]
+//        public async Task TempFilter()
+//        {
+//            const string filterName = "tempFilter";
 
-            CancellationToken token = this.TestToken;
+// string filterImage = Context.Current.TempFilterImage.Expect(
+//                () => new ArgumentException("tempFilterImage parameter is required for TempFilter test"));
+//            string sensorImage = Context.Current.TempSensorImage.Expect(
+//                () => new ArgumentException("tempSensorImage parameter is required for TempFilter test"));
 
-            EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
-                builder =>
-                {
-                    builder.AddModule(SensorName, sensorImage)
-                        .WithEnvironment(new[] { ("MessageCount", "1") });
-                    builder.AddModule(filterName, filterImage)
-                        .WithEnvironment(new[] { ("TemperatureThreshold", "19") });
-                    builder.GetModule(ModuleName.EdgeHub)
-                        .WithDesiredProperties(new Dictionary<string, object>
-                        {
-                            ["routes"] = new
-                            {
-                                TempFilterToCloud = $"FROM /messages/modules/{filterName}/outputs/alertOutput INTO $upstream",
-                                TempSensorToTempFilter = $"FROM /messages/modules/{SensorName}/outputs/temperatureOutput INTO BrokeredEndpoint('/modules/{filterName}/inputs/input1')"
-                            }
-                        } );
-                },
-                token);
+// CancellationToken token = this.TestToken;
 
-            EdgeModule filter = deployment.Modules[filterName];
-            await filter.WaitForEventsReceivedAsync(deployment.StartTime, token);
-        }
+// EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
+//                builder =>
+//                {
+//                    builder.AddModule(SensorName, sensorImage)
+//                        .WithEnvironment(new[] { ("MessageCount", "1") });
+//                    builder.AddModule(filterName, filterImage)
+//                        .WithEnvironment(new[] { ("TemperatureThreshold", "19") });
+//                    builder.GetModule(ModuleName.EdgeHub)
+//                        .WithDesiredProperties(new Dictionary<string, object>
+//                        {
+//                            ["routes"] = new
+//                            {
+//                                TempFilterToCloud = $"FROM /messages/modules/{filterName}/outputs/alertOutput INTO $upstream",
+//                                TempSensorToTempFilter = $"FROM /messages/modules/{SensorName}/outputs/temperatureOutput INTO BrokeredEndpoint('/modules/{filterName}/inputs/input1')"
+//                            }
+//                        } );
+//                },
+//                token);
 
-        [Test]
-        // Test Temperature Filter Function: https://docs.microsoft.com/en-us/azure/iot-edge/tutorial-deploy-function
-        public async Task TempFilterFunc()
-        {
-            if (OsPlatform.IsArm() && OsPlatform.Is64Bit())
-            {
-                Assert.Ignore("TempFilterFunc is disabled for arm64 because azureiotedge-functions-filter does not exist for arm64");
-            }
+// EdgeModule filter = deployment.Modules[filterName];
+//            await filter.WaitForEventsReceivedAsync(deployment.StartTime, token);
+//        }
 
-            const string filterFuncName = "tempFilterFunctions";
+// [Test]
+//        // Test Temperature Filter Function: https://docs.microsoft.com/en-us/azure/iot-edge/tutorial-deploy-function
+//        public async Task TempFilterFunc()
+//        {
+//            if (OsPlatform.IsArm() && OsPlatform.Is64Bit())
+//            {
+//                Assert.Ignore("TempFilterFunc is disabled for arm64 because azureiotedge-functions-filter does not exist for arm64");
+//            }
 
-            // Azure Function Name: EdgeHubTrigger-CSharp
-            string filterFuncImage = Context.Current.TempFilterFuncImage.Expect(
-                () => new ArgumentException("tempFilterFuncImage parameter is required for TempFilterFunc test"));
-            string sensorImage = Context.Current.TempSensorImage.Expect(
-                () => new ArgumentException("tempSensorImage parameter is required for TempFilterFunc test"));
+// const string filterFuncName = "tempFilterFunctions";
 
-            CancellationToken token = this.TestToken;
+// // Azure Function Name: EdgeHubTrigger-CSharp
+//            string filterFuncImage = Context.Current.TempFilterFuncImage.Expect(
+//                () => new ArgumentException("tempFilterFuncImage parameter is required for TempFilterFunc test"));
+//            string sensorImage = Context.Current.TempSensorImage.Expect(
+//                () => new ArgumentException("tempSensorImage parameter is required for TempFilterFunc test"));
 
-            EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
-                builder =>
-                {
-                    builder.AddModule(SensorName, sensorImage)
-                        .WithEnvironment(new[] { ("MessageCount", "1") });
-                    builder.AddModule(filterFuncName, filterFuncImage)
-                        .WithEnvironment(new[] { ("AZURE_FUNCTIONS_ENVIRONMENT", "Development") });
-                    builder.GetModule(ModuleName.EdgeHub)
-                        .WithDesiredProperties(new Dictionary<string, object>
-                        {
-                            ["routes"] = new
-                            {
-                                TempFilterFunctionsToCloud = $"FROM /messages/modules/{filterFuncName}/outputs/output1 INTO $upstream",
-                                TempSensorToTempFilter = $"FROM /messages/modules/{SensorName}/outputs/temperatureOutput INTO BrokeredEndpoint('/modules/{filterFuncName}/inputs/input1')"
-                            }
-                        });
-                },
-                token);
+// CancellationToken token = this.TestToken;
 
-            EdgeModule filter = deployment.Modules[filterFuncName];
-            await filter.WaitForEventsReceivedAsync(deployment.StartTime, token);
-        }
+// EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
+//                builder =>
+//                {
+//                    builder.AddModule(SensorName, sensorImage)
+//                        .WithEnvironment(new[] { ("MessageCount", "1") });
+//                    builder.AddModule(filterFuncName, filterFuncImage)
+//                        .WithEnvironment(new[] { ("AZURE_FUNCTIONS_ENVIRONMENT", "Development") });
+//                    builder.GetModule(ModuleName.EdgeHub)
+//                        .WithDesiredProperties(new Dictionary<string, object>
+//                        {
+//                            ["routes"] = new
+//                            {
+//                                TempFilterFunctionsToCloud = $"FROM /messages/modules/{filterFuncName}/outputs/output1 INTO $upstream",
+//                                TempSensorToTempFilter = $"FROM /messages/modules/{SensorName}/outputs/temperatureOutput INTO BrokeredEndpoint('/modules/{filterFuncName}/inputs/input1')"
+//                            }
+//                        });
+//                },
+//                token);
 
-        [Test]
-        [Category("CentOsSafe")]
-        public async Task ModuleToModuleDirectMethod(
-            [Values] Protocol protocol)
-        {
-            if (OsPlatform.IsWindows() && (protocol == Protocol.AmqpWs || protocol == Protocol.MqttWs))
-            {
-                Assert.Ignore("Module-to-module direct methods don't work over WebSocket on Windows");
-            }
+// EdgeModule filter = deployment.Modules[filterFuncName];
+//            await filter.WaitForEventsReceivedAsync(deployment.StartTime, token);
+//        }
 
-            string senderImage = Context.Current.MethodSenderImage.Expect(() => new InvalidOperationException("Missing Direct Method Sender image"));
-            string receiverImage = Context.Current.MethodReceiverImage.Expect(() => new InvalidOperationException("Missing Direct Method Receiver image"));
-            string methodSender = $"methodSender-{protocol.ToString()}";
-            string methodReceiver = $"methodReceiver-{protocol.ToString()}";
+// [Test]
+//        [Category("CentOsSafe")]
+//        public async Task ModuleToModuleDirectMethod(
+//            [Values] Protocol protocol)
+//        {
+//            if (OsPlatform.IsWindows() && (protocol == Protocol.AmqpWs || protocol == Protocol.MqttWs))
+//            {
+//                Assert.Ignore("Module-to-module direct methods don't work over WebSocket on Windows");
+//            }
 
-            CancellationToken token = this.TestToken;
+// string senderImage = Context.Current.MethodSenderImage.Expect(() => new InvalidOperationException("Missing Direct Method Sender image"));
+//            string receiverImage = Context.Current.MethodReceiverImage.Expect(() => new InvalidOperationException("Missing Direct Method Receiver image"));
+//            string methodSender = $"methodSender-{protocol.ToString()}";
+//            string methodReceiver = $"methodReceiver-{protocol.ToString()}";
 
-            EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
-                builder =>
-                {
-                    string clientTransport = protocol.ToTransportType().ToString();
+// CancellationToken token = this.TestToken;
 
-                    builder.AddModule(methodSender, senderImage)
-                        .WithEnvironment(
-                            new[]
-                            {
-                                ("TransportType", clientTransport),
-                                ("TargetModuleId", methodReceiver)
-                            });
-                    builder.AddModule(methodReceiver, receiverImage)
-                        .WithEnvironment(new[] { ("ClientTransportType", clientTransport) });
-                },
-                token);
+// EdgeDeployment deployment = await this.runtime.DeployConfigurationAsync(
+//                builder =>
+//                {
+//                    string clientTransport = protocol.ToTransportType().ToString();
 
-            EdgeModule sender = deployment.Modules[methodSender];
-            await sender.WaitForEventsReceivedAsync(deployment.StartTime, token);
-        }
-    }
-}
+// builder.AddModule(methodSender, senderImage)
+//                        .WithEnvironment(
+//                            new[]
+//                            {
+//                                ("TransportType", clientTransport),
+//                                ("TargetModuleId", methodReceiver)
+//                            });
+//                    builder.AddModule(methodReceiver, receiverImage)
+//                        .WithEnvironment(new[] { ("ClientTransportType", clientTransport) });
+//                },
+//                token);
+
+// EdgeModule sender = deployment.Modules[methodSender];
+//            await sender.WaitForEventsReceivedAsync(deployment.StartTime, token);
+//        }
+//    }
+// }
