+++ test/Microsoft.Azure.Devices.Edge.Test/Provisioning.cs
@@ -1,127 +1,128 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
-{
-    using System;
-    using System.Security.Cryptography;
-    using System.Text;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Devices.Edge.Test.Common;
-    using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
-    using Microsoft.Azure.Devices.Edge.Test.Helpers;
-    using Microsoft.Azure.Devices.Edge.Util;
-    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
-    using NUnit.Framework;
 
-    [EndToEnd]
-    public class Provisioning : DeviceProvisioningFixture
-    {
-        protected readonly IotHub iotHub;
+// namespace Microsoft.Azure.Devices.Edge.Test
+// {
+//    using System;
+//    using System.Security.Cryptography;
+//    using System.Text;
+//    using System.Threading;
+//    using System.Threading.Tasks;
+//    using Microsoft.Azure.Devices.Edge.Test.Common;
+//    using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
+//    using Microsoft.Azure.Devices.Edge.Test.Helpers;
+//    using Microsoft.Azure.Devices.Edge.Util;
+//    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
+//    using NUnit.Framework;
 
-        public Provisioning()
-        {
-            this.iotHub = new IotHub(
-                Context.Current.ConnectionString,
-                Context.Current.EventHubEndpoint,
-                Context.Current.Proxy);
-        }
+// [EndToEnd]
+//    public class Provisioning : DeviceProvisioningFixture
+//    {
+//        protected readonly IotHub iotHub;
 
-        string DeriveDeviceKey(byte[] groupKey, string registrationId)
-        {
-            using (var hmac = new HMACSHA256(groupKey))
-            {
-                return Convert.ToBase64String(hmac.ComputeHash(Encoding.UTF8.GetBytes(registrationId)));
-            }
-        }
+// public Provisioning()
+//        {
+//            this.iotHub = new IotHub(
+//                Context.Current.ConnectionString,
+//                Context.Current.EventHubEndpoint,
+//                Context.Current.Proxy);
+//        }
 
-        [Test]
-        [Category("CentOsSafe")]
-        public async Task DpsSymmetricKey()
-        {
-            string idScope = Context.Current.DpsIdScope.Expect(() => new InvalidOperationException("Missing DPS ID scope"));
-            string groupKey = Context.Current.DpsGroupKey.Expect(() => new InvalidOperationException("Missing DPS enrollment group key"));
-            string registrationId = DeviceId.Current.Generate();
+// string DeriveDeviceKey(byte[] groupKey, string registrationId)
+//        {
+//            using (var hmac = new HMACSHA256(groupKey))
+//            {
+//                return Convert.ToBase64String(hmac.ComputeHash(Encoding.UTF8.GetBytes(registrationId)));
+//            }
+//        }
 
-            string deviceKey = this.DeriveDeviceKey(Convert.FromBase64String(groupKey), registrationId);
+// [Test]
+//        [Category("CentOsSafe")]
+//        public async Task DpsSymmetricKey()
+//        {
+//            string idScope = Context.Current.DpsIdScope.Expect(() => new InvalidOperationException("Missing DPS ID scope"));
+//            string groupKey = Context.Current.DpsGroupKey.Expect(() => new InvalidOperationException("Missing DPS enrollment group key"));
+//            string registrationId = DeviceId.Current.Generate();
 
-            CancellationToken token = this.TestToken;
+// string deviceKey = this.DeriveDeviceKey(Convert.FromBase64String(groupKey), registrationId);
 
-            await this.daemon.ConfigureAsync(
-                config =>
-                {
-                    config.SetDpsSymmetricKey(idScope, registrationId, deviceKey);
-                    config.Update();
-                    return Task.FromResult((
-                        "with DPS symmetric key attestation for '{Identity}'",
-                        new object[] { registrationId }));
-                },
-                token);
+// CancellationToken token = this.TestToken;
 
-            await this.daemon.WaitForStatusAsync(EdgeDaemonStatus.Running, token);
+// await this.daemon.ConfigureAsync(
+//                config =>
+//                {
+//                    config.SetDpsSymmetricKey(idScope, registrationId, deviceKey);
+//                    config.Update();
+//                    return Task.FromResult((
+//                        "with DPS symmetric key attestation for '{Identity}'",
+//                        new object[] { registrationId }));
+//                },
+//                token);
 
-            var agent = new EdgeAgent(registrationId, this.iotHub);
-            await agent.WaitForStatusAsync(EdgeModuleStatus.Running, token);
-            await agent.PingAsync(token);
+// await this.daemon.WaitForStatusAsync(EdgeDaemonStatus.Running, token);
 
-            Option<EdgeDevice> device = await EdgeDevice.GetIdentityAsync(
-                registrationId,
-                this.iotHub,
-                token,
-                takeOwnership: true);
+// var agent = new EdgeAgent(registrationId, this.iotHub);
+//            await agent.WaitForStatusAsync(EdgeModuleStatus.Running, token);
+//            await agent.PingAsync(token);
 
-            Context.Current.DeleteList.TryAdd(
-                registrationId,
-                device.Expect(() => new InvalidOperationException(
-                    $"Device '{registrationId}' should have been created by DPS, but was not found in '{this.iotHub.Hostname}'")));
-        }
+// Option<EdgeDevice> device = await EdgeDevice.GetIdentityAsync(
+//                registrationId,
+//                this.iotHub,
+//                token,
+//                takeOwnership: true);
 
-        [Test]
-        public async Task DpsX509()
-        {
-            (string, string, string) rootCa =
-                        Context.Current.RootCaKeys.Expect(() => new InvalidOperationException("Missing root CA keys"));
-            string caCertScriptPath =
-                        Context.Current.CaCertScriptPath.Expect(() => new InvalidOperationException("Missing CA cert script path"));
-            string idScope = Context.Current.DpsIdScope.Expect(() => new InvalidOperationException("Missing DPS ID scope"));
-            string registrationId = DeviceId.Current.Generate();
+// Context.Current.DeleteList.TryAdd(
+//                registrationId,
+//                device.Expect(() => new InvalidOperationException(
+//                    $"Device '{registrationId}' should have been created by DPS, but was not found in '{this.iotHub.Hostname}'")));
+//        }
 
-            CancellationToken token = this.TestToken;
+// [Test]
+//        public async Task DpsX509()
+//        {
+//            (string, string, string) rootCa =
+//                        Context.Current.RootCaKeys.Expect(() => new InvalidOperationException("Missing root CA keys"));
+//            string caCertScriptPath =
+//                        Context.Current.CaCertScriptPath.Expect(() => new InvalidOperationException("Missing CA cert script path"));
+//            string idScope = Context.Current.DpsIdScope.Expect(() => new InvalidOperationException("Missing DPS ID scope"));
+//            string registrationId = DeviceId.Current.Generate();
 
-            CertificateAuthority ca = await CertificateAuthority.CreateAsync(
-                registrationId,
-                rootCa,
-                caCertScriptPath,
-                token);
+// CancellationToken token = this.TestToken;
 
-            IdCertificates idCert = await ca.GenerateIdentityCertificatesAsync(registrationId, token);
+// CertificateAuthority ca = await CertificateAuthority.CreateAsync(
+//                registrationId,
+//                rootCa,
+//                caCertScriptPath,
+//                token);
 
-            await this.daemon.ConfigureAsync(
-                config =>
-                {
-                    config.SetDpsX509(idScope, registrationId, idCert);
-                    config.Update();
-                    return Task.FromResult((
-                        "with DPS X509 attestation for '{Identity}'",
-                        new object[] { registrationId }));
-                },
-                token);
+// IdCertificates idCert = await ca.GenerateIdentityCertificatesAsync(registrationId, token);
 
-            await this.daemon.WaitForStatusAsync(EdgeDaemonStatus.Running, token);
+// await this.daemon.ConfigureAsync(
+//                config =>
+//                {
+//                    config.SetDpsX509(idScope, registrationId, idCert);
+//                    config.Update();
+//                    return Task.FromResult((
+//                        "with DPS X509 attestation for '{Identity}'",
+//                        new object[] { registrationId }));
+//                },
+//                token);
 
-            var agent = new EdgeAgent(registrationId, this.iotHub);
-            await agent.WaitForStatusAsync(EdgeModuleStatus.Running, token);
-            await agent.PingAsync(token);
+// await this.daemon.WaitForStatusAsync(EdgeDaemonStatus.Running, token);
 
-            Option<EdgeDevice> device = await EdgeDevice.GetIdentityAsync(
-                registrationId,
-                this.iotHub,
-                token,
-                takeOwnership: true);
+// var agent = new EdgeAgent(registrationId, this.iotHub);
+//            await agent.WaitForStatusAsync(EdgeModuleStatus.Running, token);
+//            await agent.PingAsync(token);
 
-            Context.Current.DeleteList.TryAdd(
-                registrationId,
-                device.Expect(() => new InvalidOperationException(
-                    $"Device '{registrationId}' should have been created by DPS, but was not found in '{this.iotHub.Hostname}'")));
-        }
-    }
-}
+// Option<EdgeDevice> device = await EdgeDevice.GetIdentityAsync(
+//                registrationId,
+//                this.iotHub,
+//                token,
+//                takeOwnership: true);
+
+// Context.Current.DeleteList.TryAdd(
+//                registrationId,
+//                device.Expect(() => new InvalidOperationException(
+//                    $"Device '{registrationId}' should have been created by DPS, but was not found in '{this.iotHub.Hostname}'")));
+//        }
+//    }
+// }
