+++ test/Microsoft.Azure.Devices.Edge.Test/SetupFixture.cs
@@ -1,106 +1,107 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
-{
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Net;
-    using System.Text;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Devices.Edge.Test.Common;
-    using Microsoft.Azure.Devices.Edge.Test.Helpers;
-    using Microsoft.Azure.Devices.Edge.Util;
-    using NUnit.Framework;
-    using Serilog;
-    using Serilog.Events;
 
-    [SetUpFixture]
-    public class SetupFixture
-    {
-        IEdgeDaemon daemon;
+// namespace Microsoft.Azure.Devices.Edge.Test
+// {
+//    using System.Collections.Generic;
+//    using System.Linq;
+//    using System.Net;
+//    using System.Text;
+//    using System.Threading;
+//    using System.Threading.Tasks;
+//    using Microsoft.Azure.Devices.Edge.Test.Common;
+//    using Microsoft.Azure.Devices.Edge.Test.Helpers;
+//    using Microsoft.Azure.Devices.Edge.Util;
+//    using NUnit.Framework;
+//    using Serilog;
+//    using Serilog.Events;
 
-        [OneTimeSetUp]
-        public async Task BeforeAllAsync()
-        {
-            using var cts = new CancellationTokenSource(Context.Current.SetupTimeout);
-            CancellationToken token = cts.Token;
-            Option<Registry> bootstrapRegistry = Option.Maybe(Context.Current.Registries.First());
+// [SetUpFixture]
+//    public class SetupFixture
+//    {
+//        IEdgeDaemon daemon;
 
-            this.daemon = await OsPlatform.Current.CreateEdgeDaemonAsync(
-                Context.Current.InstallerPath,
-                Context.Current.EdgeAgentBootstrapImage,
-                bootstrapRegistry,
-                token);
+// [OneTimeSetUp]
+//        public async Task BeforeAllAsync()
+//        {
+//            using var cts = new CancellationTokenSource(Context.Current.SetupTimeout);
+//            CancellationToken token = cts.Token;
+//            Option<Registry> bootstrapRegistry = Option.Maybe(Context.Current.Registries.First());
 
-            await Profiler.Run(
-                async () =>
-                {
-                    // Set up logging
-                    LogEventLevel consoleLevel = Context.Current.Verbose
-                        ? LogEventLevel.Verbose
-                        : LogEventLevel.Information;
-                    var loggerConfig = new LoggerConfiguration()
-                        .MinimumLevel.Verbose()
-                        .WriteTo.NUnit(consoleLevel);
-                    Context.Current.LogFile.ForEach(f => loggerConfig.WriteTo.File(f));
-                    Log.Logger = loggerConfig.CreateLogger();
+// this.daemon = await OsPlatform.Current.CreateEdgeDaemonAsync(
+//                Context.Current.InstallerPath,
+//                Context.Current.EdgeAgentBootstrapImage,
+//                bootstrapRegistry,
+//                token);
 
-                    // Install IoT Edge, and do some basic configuration
-                    await this.daemon.UninstallAsync(token);
-                    await this.daemon.InstallAsync(Context.Current.PackagePath, Context.Current.Proxy, token);
+// await Profiler.Run(
+//                async () =>
+//                {
+//                    // Set up logging
+//                    LogEventLevel consoleLevel = Context.Current.Verbose
+//                        ? LogEventLevel.Verbose
+//                        : LogEventLevel.Information;
+//                    var loggerConfig = new LoggerConfiguration()
+//                        .MinimumLevel.Verbose()
+//                        .WriteTo.NUnit(consoleLevel);
+//                    Context.Current.LogFile.ForEach(f => loggerConfig.WriteTo.File(f));
+//                    Log.Logger = loggerConfig.CreateLogger();
 
-                    await this.daemon.ConfigureAsync(
-                        config =>
-                        {
-                            var msgBuilder = new StringBuilder();
-                            var props = new List<object>();
+// // Install IoT Edge, and do some basic configuration
+//                    await this.daemon.UninstallAsync(token);
+//                    await this.daemon.InstallAsync(Context.Current.PackagePath, Context.Current.Proxy, token);
 
-                            string hostname = Dns.GetHostName();
-                            config.SetDeviceHostname(hostname);
-                            msgBuilder.Append("with hostname '{hostname}'");
-                            props.Add(hostname);
+// await this.daemon.ConfigureAsync(
+//                        config =>
+//                        {
+//                            var msgBuilder = new StringBuilder();
+//                            var props = new List<object>();
 
-                            Context.Current.ParentHostname.ForEach(parentHostname =>
-                            {
-                                config.SetParentHostname(parentHostname);
-                                msgBuilder.AppendLine(", parent hostname '{parentHostname}'");
-                                props.Add(parentHostname);
-                            });
+// string hostname = Dns.GetHostName();
+//                            config.SetDeviceHostname(hostname);
+//                            msgBuilder.Append("with hostname '{hostname}'");
+//                            props.Add(hostname);
 
-                            Context.Current.Proxy.ForEach(proxy =>
-                            {
-                                config.AddHttpsProxy(proxy);
-                                msgBuilder.AppendLine(", proxy '{ProxyUri}'");
-                                props.Add(proxy.ToString());
-                            });
+// Context.Current.ParentHostname.ForEach(parentHostname =>
+//                            {
+//                                config.SetParentHostname(parentHostname);
+//                                msgBuilder.AppendLine(", parent hostname '{parentHostname}'");
+//                                props.Add(parentHostname);
+//                            });
 
-                            config.Update();
+// Context.Current.Proxy.ForEach(proxy =>
+//                            {
+//                                config.AddHttpsProxy(proxy);
+//                                msgBuilder.AppendLine(", proxy '{ProxyUri}'");
+//                                props.Add(proxy.ToString());
+//                            });
 
-                            return Task.FromResult((msgBuilder.ToString(), props.ToArray()));
-                        },
-                        token,
-                        restart: false);
-                },
-                "Completed end-to-end test setup");
-        }
+// config.Update();
 
-        [OneTimeTearDown]
-        public Task AfterAllAsync() => TryFinally.DoAsync(
-            () => Profiler.Run(
-                async () =>
-                {
-                    using var cts = new CancellationTokenSource(Context.Current.TeardownTimeout);
-                    CancellationToken token = cts.Token;
-                    await this.daemon.StopAsync(token);
-                    foreach (EdgeDevice device in Context.Current.DeleteList.Values)
-                    {
-                        await device.MaybeDeleteIdentityAsync(token);
-                    }
-                },
-                "Completed end-to-end test teardown"),
-            () =>
-            {
-                Log.CloseAndFlush();
-            });
-    }
-}
+// return Task.FromResult((msgBuilder.ToString(), props.ToArray()));
+//                        },
+//                        token,
+//                        restart: false);
+//                },
+//                "Completed end-to-end test setup");
+//        }
+
+// [OneTimeTearDown]
+//        public Task AfterAllAsync() => TryFinally.DoAsync(
+//            () => Profiler.Run(
+//                async () =>
+//                {
+//                    using var cts = new CancellationTokenSource(Context.Current.TeardownTimeout);
+//                    CancellationToken token = cts.Token;
+//                    await this.daemon.StopAsync(token);
+//                    foreach (EdgeDevice device in Context.Current.DeleteList.Values)
+//                    {
+//                        await device.MaybeDeleteIdentityAsync(token);
+//                    }
+//                },
+//                "Completed end-to-end test teardown"),
+//            () =>
+//            {
+//                Log.CloseAndFlush();
+//            });
+//    }
+// }
