+++ test/Microsoft.Azure.Devices.Edge.Test/X509Device.cs
@@ -1,52 +1,53 @@
 // Copyright (c) Microsoft. All rights reserved.
-namespace Microsoft.Azure.Devices.Edge.Test
-{
-    using System;
-    using System.Threading;
-    using System.Threading.Tasks;
-    using Microsoft.Azure.Devices.Edge.Test.Common;
-    using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
-    using Microsoft.Azure.Devices.Edge.Test.Common.Config;
-    using Microsoft.Azure.Devices.Edge.Test.Helpers;
-    using Microsoft.Azure.Devices.Edge.Util;
-    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
-    using NUnit.Framework;
 
-    [EndToEnd]
-    class X509Device : X509ManualProvisioningFixture
-    {
-        [Test]
-        public async Task X509ManualProvision()
-        {
-            CancellationToken token = this.TestToken;
+// namespace Microsoft.Azure.Devices.Edge.Test
+// {
+//    using System;
+//    using System.Threading;
+//    using System.Threading.Tasks;
+//    using Microsoft.Azure.Devices.Edge.Test.Common;
+//    using Microsoft.Azure.Devices.Edge.Test.Common.Certs;
+//    using Microsoft.Azure.Devices.Edge.Test.Common.Config;
+//    using Microsoft.Azure.Devices.Edge.Test.Helpers;
+//    using Microsoft.Azure.Devices.Edge.Util;
+//    using Microsoft.Azure.Devices.Edge.Util.Test.Common.NUnit;
+//    using NUnit.Framework;
 
-            await this.runtime.DeployConfigurationAsync(token);
+// [EndToEnd]
+//    class X509Device : X509ManualProvisioningFixture
+//    {
+//        [Test]
+//        public async Task X509ManualProvision()
+//        {
+//            CancellationToken token = this.TestToken;
 
-            string leafDeviceId = DeviceId.Current.Generate();
+// await this.runtime.DeployConfigurationAsync(token);
 
-            var leaf = await LeafDevice.CreateAsync(
-                leafDeviceId,
-                Protocol.Amqp,
-                AuthenticationType.Sas,
-                Option.Some(this.runtime.DeviceId),
-                false,
-                CertificateAuthority.GetQuickstart(),
-                this.iotHub,
-                token,
-                Option.None<string>());
+// string leafDeviceId = DeviceId.Current.Generate();
 
-            await TryFinally.DoAsync(
-                async () =>
-                {
-                    DateTime seekTime = DateTime.Now;
-                    await leaf.SendEventAsync(token);
-                    await leaf.WaitForEventsReceivedAsync(seekTime, token);
-                    await leaf.InvokeDirectMethodAsync(token);
-                },
-                async () =>
-                {
-                    await leaf.DeleteIdentityAsync(token);
-                });
-        }
-    }
-}
+// var leaf = await LeafDevice.CreateAsync(
+//                leafDeviceId,
+//                Protocol.Amqp,
+//                AuthenticationType.Sas,
+//                Option.Some(this.runtime.DeviceId),
+//                false,
+//                CertificateAuthority.GetQuickstart(),
+//                this.iotHub,
+//                token,
+//                Option.None<string>());
+
+// await TryFinally.DoAsync(
+//                async () =>
+//                {
+//                    DateTime seekTime = DateTime.Now;
+//                    await leaf.SendEventAsync(token);
+//                    await leaf.WaitForEventsReceivedAsync(seekTime, token);
+//                    await leaf.InvokeDirectMethodAsync(token);
+//                },
+//                async () =>
+//                {
+//                    await leaf.DeleteIdentityAsync(token);
+//                });
+//        }
+//    }
+// }
